Add brillo_NVRAM test.
This is intended to be a Brillo PTS sanity test for the Brillo
Access-Controlled NVRAM hardware requirements. This test should fail if
the HAL module is the software-only 'testing' module.

This test relies on the existence of /data/nativetest binaries which
means it will not pass on a production image.

BUG=b:27320510
TEST=build, run on brilloemulator

Change-Id: Iaf886b745099ed8989e151516565f5e8d94cfdc4
Reviewed-on: https://chromium-review.googlesource.com/333140
Commit-Ready: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>