[Autotest] add touch_TouchscreenZoom test
Add an autotest to confirm that touchscreen zooms occur in
Chrome as expected.  Use touch playback library to play
a touchscreen pinch zoom in and then a pinch zoom out.
Confirm that the the Chrome webpage zooms in and then out.

Only on Samus for now.

TEST=ran many times on samus.  saw no flake
BUG=chromium:537472

Change-Id: I6a604acf92eeb55d3c027aea15907b3801b49bb2
Reviewed-on: https://chromium-review.googlesource.com/333401
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Tien Chang <tienchang@chromium.org>