[autotest] Rename the exception for Verify errors.
This renames the exceptions to be raised when Verifiers fail.
AutotestHostVerifyError is renamed to AutoservVerifyError, and
AutotestVerifyDependencyError becomes AutoservVerifyDependencyError.
The change is being made for naming consistency with existing exception
classes, and with some upcoming names to be added with the new
repair framework.

BUG=None
TEST=unit tests; manually run verify to provoke selected errors.

Change-Id: I7c4520c8f87189aaf847c7391946872eddfcdd66
Reviewed-on: https://chromium-review.googlesource.com/333661
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>