Autotest: reboot DUTs when they are moved from shard to master.
A special task REPAIR is triggered to force rebooting on the shard DUTs
when a shard is deleted, in order to make sure the DUTs are still of
ready status and own testing logs in master DB. However, the REPAIR job
won't support reboot in a short time.

This CL triggers a reboot test with highest priority on all DUTs that will
be moved from shard to master. The procedure for deleting a shard is:
1. unlock all related DUTs of this shard.
2. delete any shard information in master DB.
3. trigger a reboot test with highest priority, to make sure that this
test runs firstly after the DUTs are unlocked.
4. unlock these DUTs.

BUG=chromium:499865
TEST=Configer a cbf master and a cbf shard. Set several tasks on master, one
is running, others are pending. Ran 'atest shard delete ***' on the master
to make sure:
    * The DUTs beloging to the shard are locked.
    * The shard is deleted
    * A reboot with highest priority is triggered.
    * The DUTs are unlocked.
    * The reboot test is ran first, all other pending tasks are queued still.
    * After reboot test is finished, other pending tasks will continue
      to run on the master.
Ran site_rpc_interface_unittest locally.

Change-Id: I2b348e520c0f67bec5b4b1c89c75ad41e86c72a2
Reviewed-on: https://chromium-review.googlesource.com/334434
Commit-Ready: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>