firmware_EventLog: Test for hardware watchdog event
This patch adds a check for the watchdog timeout event to the
firmware_EventLog test. Skipping x86 devices for now which don't seem to
report this event (at least on my old Falco).

Also change all the TestError in this test to TestFail which is actually
the correct exception to throw for cases where the board doesn't meet
expectations (as opposed to cases where the test logic itself ran into
an unexpected condition somehow).

BUG=None
TEST=Ran on Oak.

Change-Id: Ie00e471b5cad5048ca76a1bf13b5f4d4b9946696
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/334744
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>