Autotest: Make Devserver retry ssh connection, and Add stats for ssh failures.
When trying to SSH to a devserver, this SSH connection could be aborted
by the devserver itself, which causes staging failure.

This CL add 'retry' to all ssh connections to devserver, and add stats
for ssh failures: ssh_dev_server_failure, and ssh_tunnel_failure, to
estimate the stabality of ssh-based approaches.

BUG=chromium:597680
TEST=Ran dev_server_unittest to ensure error.CmdError causes retry. Turn
on the enable_ssh switch to test that ssh to devserver & servo still works.

Change-Id: I4b9e6092d8989fb2ecb770367c5869042e8e5740
Reviewed-on: https://chromium-review.googlesource.com/334770
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Xixuan Wu <xixuan@chromium.org>