logging_GenerateCrashFiles: Add SSH timeout to crash command
logging_GenerateCrashFiles.KERNEL currently takes forever (or, more
precisely, 15 minutes) to complete, because the crashing SSH command
never completes from the server's point of view and it waits the whole
-o ServerAliveInterval=900 seconds (as hardcoded in ssh_host.py) until
it tries to continue the test. Let's add an explicit (but ignored)
timeout to make this less tedious to run.

BUG=None
TEST=Ran on Oak.

Change-Id: I8a5dd1e06197329096dcd8d9d3326a4e7ea839e8
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/335572
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>