Autotest: Add specified exception type to Retry for dev_server and servo
After adding ssh feature to dev_server and servo, we need to do retry
when meeting error.CmdError (ssh error). However a specified error type
needs to raised after retry, in order to be caught by its caller.

For example, a caller needs to call dev_server.get_control_file(). The ideal
case is we retry ssh connections to dev_server if it fails. When retry
is finished, we need it to raise a DevServerException, which is
currently caught by the caller.

This CL adds a specified exception to raise in retry, and also
refactors the error msg raising processes.

BUG=chromium:598496
TEST=Checked dev_server and servo raise the corresponding error types in
local autotest instance. Ran dev_server_unittest.

Change-Id: I53ad52b2c397beeb922d75539b63eaf9b6632c86
Reviewed-on: https://chromium-review.googlesource.com/335901
Commit-Ready: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>