[autotest] Switch the default FAFT suites to test the RO (and RW) firmware
The original FAFT suites (faft_bios and faft_ec) pick the RO firmware
from a stable build and the RW from the top of the firmware branch.

This change switches them to test the RO (and also RW) firmware from
the top of the firmware branch.

The original combination (RW from the top of the firmware branch) will
be added back in a following change.

BUG=chromium:575394
TEST=None

Change-Id: I6da2b8bcb00bb3aad1e0bb05e95a610a294024c2
Reviewed-on: https://chromium-review.googlesource.com/336877
Commit-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>