[autotest] Update _get_devices to catch a general Exception for fastboot.
The latest update to getting the device serials via fastboot handle the
situation where fastboot does not return serials for TCP.  There's a
check to see if the device is in fastboot mode with an expected timeout
but the try/catch looks for the wrong exception.  I've switched out the
exception to be a general Exception with a comment explaining why.

BUG=None
TEST=locally on moblab

Change-Id: I47acf3d805e120aa471767ab0c8557c23c669723
Reviewed-on: https://chromium-review.googlesource.com/338871
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>