Disable telemetry GPU tests in graphics suite
This patch disables the telemetry GPU tests to
be run as part of "graphics" suite.

Telemetry tests are more browser specific and have
to be executed as explained in link:
https://www.chromium.org/developers/telemetry/running-telemetry-on-chrome-os

Also since its browser source code linked, its not only
dependant on the browser code linked during execution but also
remains broken till issue (if any) is fixed in the browser.

However these tests are part of "suite:graphics_browser",
which is more apt as per the functionality they verify.

Have removed suite:graphics from generate_controlfiles.py.
And rerun the scripts.

BUG=chrome-os-partner:46233
TEST=None

Change-Id: Iae1153f08a536c5ba1b02ba1723011fb8caaa982
Signed-off-by: Shirish S <shirish.s@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/338717
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>