Store temporary perf data in local autotest directory.
Problem: Longevity_Tracker.py stores the temporary perf data in the system
/tmp/ directory during the test run. The problem with this is that, if the
device crashes during the test, the data in the /tmp/ directory is often lost.

Solution: Create the temporary perf file in the /usr/local/autotest/
directory. When the test exits, copy the temporary file to 'perf.csv' stored
in the ../results directory, where AutoTest will upload it with the rest of
the test results.

BUG=chromium:603226
TEST=test_that <IP> f:.*longevity_Tracker/control.longevity_2

Change-Id: I20a2e6717c67002e55822aad9ffed14abba557d3
Reviewed-on: https://chromium-review.googlesource.com/339331
Commit-Ready: Scott Cunningham <scunningham@chromium.org>
Tested-by: Scott Cunningham <scunningham@chromium.org>
Reviewed-by: Scott Cunningham <scunningham@chromium.org>