[autotest] Allow create_suite_job RPC to pick a devserver in desired subnet.
This change allows create_suite_job RPC to pick a devserver in the same subnet
of the duts may run tests. The assumption is that duts in the same pool and
board are in the same subnet, which should be true once ACL is enforced.

BUG=chromium:577363
TEST=run in local setup
update global config:
prefer_local_devserver: True
add couple devservers in different subnet
call run_suite, confirm the desired desverver is picked

Change-Id: I3435c13a7c5b25a429609780e0d8471a066e501d
Reviewed-on: https://chromium-review.googlesource.com/339250
Commit-Ready: Dan Shi <dshi@google.com>
Tested-by: Dan Shi <dshi@google.com>
Reviewed-by: Dan Shi <dshi@google.com>