Moving client test result parser to autotest.py
When a server test initiates a client test, it is important to reflect
the failure of the client test in wmatrix. With the current setup only
the server test result is reflected in wmatrix when tests are run with
such setup. This hides all client test failures which are only seen if
one looks at the test logs. We want to make sure that the server test
fails anytime the client test initiated by this server test does not
pass. This was currently done in hardware_StorageStress test by parsing
the client test results. I am moving this code to autotest.py so all
such tests can benefit.

BUG=None
TEST=Locally

Change-Id: I0f033b2d25fe777f1ebe7e7573ee9f24df06a53a
Reviewed-on: https://chromium-review.googlesource.com/339811
Commit-Ready: harpreet Grewal <harpreet@chromium.org>
Tested-by: harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>