[autotest]Support for packet capturing using a second AP.
Currently the wifi test context manager only deals with a single
router that also acts as the packet capturer. With the new-style
wifi cells with whirwlind, packet capturer is a separate entity.

Added a new member to the context manger 'pcap_host'. 'pcap_host'
plays the role of exclusive packet capturing on a dedicated AP.
The wifi_test_context_manger sets the capture host to the second AP
if preset or it defaults to the router. It is transparent to the
test, which of the APs is being  used for packet capturing.

While here, also changed the commmand line option to enable packet
capturing while abstracting away the details of router vs. dedicated AP.
CMDLINE_PACKET_CAPTURES = 'packet_capture'

BUG=chromium:601238
TEST=Simple connect/matfunc

Change-Id: I58f76ea5e99e700b29c87359fc18345c5db435e8
Reviewed-on: https://chromium-review.googlesource.com/340270
Commit-Ready: Bindu Mahadev <bmahadev@chromium.org>
Commit-Ready: Kris Rambish <krisr@chromium.org>
Tested-by: Bindu Mahadev <bmahadev@chromium.org>
Reviewed-by: Bindu Mahadev <bmahadev@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>