faft: Refactoring soft reset test
Moved the states_list to be part of the plankton PD device class
similar to the changes for the hard_reset() method. In addition
added try/except clauses to method calls in the soft_reset test
script.

BRANCH=none
BUG=chrome-os-partner:50178
TEST=Manual
Tested with Samus to Plankton. Verfied that soft resets are initiated
and the state transition history can be used to verify expected operation.

Change-Id: I131f6b2784c77cfdb69210250d3dd5c11fd739c9
Signed-off-by: Scott <scollyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/340248
Commit-Ready: Scott Collyer <scollyer@chromium.org>
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>