[autotest] Make ssh logging readable.
ssh commands executed by SSHHost.run() include gunk for logging the
command on the target host.  The full command (with the gunk) was
being written to the debug logs.  The gunk is useful on the target,
but not useful in the debug logs.

This change removes the gunk from the debug logs.

BUG=None
TEST=Run a verify cycle on a local autotest instance

Change-Id: I22d8c835c5a005f380f590d417b60cf4f5ce4969
Reviewed-on: https://chromium-review.googlesource.com/340359
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>