[autotest] Set brillo_StorageWriteSpeedTest emulator minimum to 1mbsp
The builder has been randomly failing the brillo_StorageWriteSpeedTest
test on the emulator.  This test is part of the PTS, and mainly only
valid on hardware.  The emulator version is just to verify that the
test's logic works, before having to wait for the test to get deployed
to hardware.  As such, set the minimum write speed down to just
1mbps (from the current value of 10mbps), just to verify that it
read in a sane value.

Bug: 28345208
Change-Id: Ib7f8cacdcebae412ed328bdc403a3faac86612ea
Reviewed-on: https://chromium-review.googlesource.com/340450
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Steve Fung <stevefung@chromium.org>
Trybot-Ready: Steve Fung <stevefung@chromium.org>
Tested-by: Steve Fung <stevefung@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>