[autotest] Add in label verifier, part 2.
This verifier will be used to ensure that the labels on the hosts will
be kept up-to-date whenever the verify process is run (which should be
daily).  Added for CrosHost (ADBHost and Testbed will be in a later cl).
Also made sure the verifier will catch any exception and log it, but not
propagate it upwards.

BUG=chromium:571088
TEST=Test that labels are properly updated when verify is run on them.
CQ-DEPEND=CL:338904

Change-Id: I03fdd49e66e3cf7c0798fa05d1984bbd9ad6ddce
Reviewed-on: https://chromium-review.googlesource.com/338879
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>