Fix ImagesAllowedForUrls and ImagesBlockedForUrls failures.
The policy_ImagesAllowedForUrls and policy_ImagesBlockedForUrls tests
began failing when Chrome started showing a broken-image gif for
blocked images, rather than a blank box.

This fix checks to see wether the displayed image is less than the
minimum expected size of the image (of some cute kittens), to determine
that the image was blocked.

The fix also updates the tests to take advantage of newer methods
in the base class.

BUG=chromium:604437
TEST=Run test_that <IP> policy_ImagesAllowedForUrls and
policy_ImagesBlockedForUrls

Change-Id: If424db9d476feae9cbadf4807737aea42a9d1433
Reviewed-on: https://chromium-review.googlesource.com/342575
Commit-Ready: Scott Cunningham <scunningham@chromium.org>
Tested-by: Scott Cunningham <scunningham@chromium.org>
Reviewed-by: Krishna Gavini <krishnargv@chromium.org>