client/cros/chameleon: add methods to get bluetooth mouse object
This patch simply adds methods to get the bluetooth mouse object
out the chameleond proxy. With the mouse object, a test could
control the bluetooth mouse to perform various mouse actions.

Specifically, two methods are added. One is to get an existing
mouse object, and the other is to create a new mouse object.
The latter is useful when we want to reconnect to the RN42
emulation kit through the serial port while the kit is
disconnected with an unknown reason sometimes.

BUG=chromium:610238
TEST=None

Change-Id: I6b7517a6450476cd23c92fc43af1a7ca2e3b7d58
Reviewed-on: https://chromium-review.googlesource.com/349898
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>