Set servo usb_mux_sel3 to dut_sees_usbkey
If servo is detected, set the bottom usb mux port on the servo to be
seen by dut. This sanity test needs to run on all hotrod devices
regardless of whether there is a servo attached or not. When servo is
present the speaker, which is tested in this sanity test, will be
attached via the bottom usb mux port so we need to make sure this port
is visible to the dut. When servo is not present the speaker will be
attached directly to 1 of the USB ports on the dut.

BUG=None
TEST=Tested locally.

Change-Id: Id663fd54c969740e9a8f3b2d07ce211244431361
Reviewed-on: https://chromium-review.googlesource.com/343915
Commit-Ready: harpreet Grewal <harpreet@chromium.org>
Tested-by: harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>