[autotest] Remove rsyslogd from testing in logging_CrashServices
The logging_CrashServices test has a list of processes that it kills
in order to test that dumps are properly gathered.  One of the
victims is rsyslogd.  Killing rsyslogd sometimes causes the DUT to
go offline.  This removes rsyslogd from the list of services to
kill, until the service can be made more robust.

BUG=chromium:611786
TEST=None

Change-Id: I0365f01733d3e80d81a91f020163cb5a9c053a76
Reviewed-on: https://chromium-review.googlesource.com/344634
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Shuqian Zhao <shuqianz@chromium.org>