Add multiprocessing setting for gs_offloader.
If the command line flag "multiprocessing" is set,  it turns on the -m option
for gsutil when gs_offloarder runs. If the flag is not set, the global config
setting "gs_offloader_multiprocessing" under the CROS section is used, which
is default to False.

BUG=chromium:590929
TEST=unit test and test on device.

Change-Id: I6580866b7691e1be29977f271dfae8d67e2e06cc
Reviewed-on: https://chromium-review.googlesource.com/344832
Commit-Ready: Michael Tang <ntang@chromium.org>
Tested-by: Michael Tang <ntang@chromium.org>
Reviewed-by: Michael Tang <ntang@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>