[autotest] Reduce polling load for dynamic suites.
Dynamic suites check for completion by polling with a query for all
finished child jobs.  At each polling cycle the only jobs of
interest would be jobs that had finished since the previous cycle,
but the query fetched all finished jobs, meaning that data for jobs
that finished early would be retransmitted many times.

The excessive amount of data transmitted for the polling query is
suspected of causing load on the RPC server which in turn causes
mysterious server hang-ups.

This changes the query to filter the unneeded data on the server
side, so that each finished job is only transmitted once.

BUG=chromium:606071
TEST=unit tests, plus push_to_prod suite on local instance

Change-Id: I27c7be51f341e3d90a3fbd92ac8f6a93699a340a
Reviewed-on: https://chromium-review.googlesource.com/344895
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>