[autotest] Remove atest cli support for specifying a test kernel.
The `atest job create` command provided a -k option for specifying
a kernel to be tested.  The option was a non-functional legacy not
applicable to Chrome OS.

This deletes the legacy atest CLI code.  The underlying RPC calls
and features made obsolete by this change will be removed in
subsequent CLs.

BUG=None
TEST=run `atest job create` in a local Autotest instance, plus unit tests

Change-Id: I882f96bf940f9de8af1ffd502e7b3f59e6fdddb5
Reviewed-on: https://chromium-review.googlesource.com/346058
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>