[autotest] Drop RPC support for specifying a test kernel.
Our legacy Autotest code included support for specifying a specific
kernel to be tested when creating a job.  That support wasn't
appropriate for Chrome OS, and didn't work.

All existing user level interfaces for specifying a kernel (CLI and
AFE) have been changed to drop this feature.  This change removes
support for the feature in the RPC calls.  Lower-level layers still
contain support that is unused, which will be removed in subsequent
CLs.

BUG=chromium:266704
TEST=Look at control file from AFE job creation page; run atest job create

Change-Id: Ifb0c6888f550f6cda98195b489ba8076cdf47e6b
Reviewed-on: https://chromium-review.googlesource.com/346570
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>