faft: USB PD: Improvements to disconnect/connect methods
PD console devices force a disconnect/connect event by forcing
drp mode to be the opposite of the current connection. Then
the disconnect and connect states are verified. However, this
approach will not work if the other device in the connection
is in dualrole mode. If both devices are in dualrole mode then
the disconnect/connect can be checked by verifying that a power
role swap occurs when the drp mode is forced.

In addition a drp_disconnect_connect() method was added to the
Plankton device so that the same method can be called from test
scripts.

BRANCH=none
BUG=chrome-os-partner:50431
TEST=Manual
Tested with Samus and Plankton using the PDConnect script.

Change-Id: If910f3992e573f0d2a25de01e965177ea32c76fc
Signed-off-by: Scott <scollyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/347446
Commit-Ready: Scott Collyer <scollyer@chromium.org>
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>