faft: Refactored PDConnect test to use pd_device objects
The existing version of this test was written prior to
developing the pd_device objects. Refactored the test to
take advantage of this encapsulation. The overall test is
now much simpler to both implement and maintain.

BRANCH=none
BUG=chrome-os-partner:53742
TEST=Manual
Tested using Samus and Plankton with workstation.

Change-Id: Id2d93fae7709d799e57b1837b5dff6a3e8148319
Signed-off-by: Scott <scollyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/347447
Commit-Ready: Scott Collyer <scollyer@chromium.org>
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>