firmware_ECKeyboardReboot: Increase wait_up time.
Currently, firmware_ECKeyboardReboot has been very flaky.  In some of my
own experiments, I've found the issue to be that the 30 seconds that we
wait for the DUT to come up is _very_ close to the time it takes for a
device to reboot and establish a link using the USB ethernet adapter.
The test often fails with the error that 'DUT is off, expected on', when
the DUT actually *is* on.

This test should really be testing whether a DUT comes up or not and not
"how quickly".  Therefore, increase the time to wait for the DUT to come
up to 60 seconds.  This won't be sufficient for a network outage, but it
at least should provide a bit more room for random delays in the
network.  No harm is done if the DUT comes up earlier.

BUG=None
BRANCH=None
TEST=Run firmware_ECKeyboardReboot on a cyan and observe failures.
Apply this patch and repeat the tests and observe no failures.

Change-Id: I3448d89ceab3dda57efcca096d6f97513eca6aa5
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/347490
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Tom Tam <waihong@google.com>