Move poll_for_condition to common_lib.
Some server test starts to use it, and there is no reason
to keep it in client.bin.site_utils from implementation
perspective.

This CL moves it to client.common_lib.site_utils, thus
test code can access to it via autotest_lib.server.utils,
instead of client.bin.utils.

Note that it is still accessible from client.bin.utils, so
no client side change is needed at the moment.

In future, probably it would be better to integrate the
function to client.common_lib.base_utils.wait_for_value() and
wait_for_value_changed(), but it is not the goal of this CL.

BUG=None
TEST=Ran locally.

Change-Id: Ic15df9d87c96ae459267cd157bbbb408abca6bfd
Reviewed-on: https://chromium-review.googlesource.com/347562
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>