bluetooth_adpater_tests: add retry function
It is observed in the field that the chameleon board might sometimes
lose the serial device temporarily. The kernel would re-detect and
reconnect the serial device. After the re-connection, it is possible
that the serial port changes as well, e.g., from /dev/ttyUSB0
to /dev/ttyUSB1. In this case, it is required to reconnect to the
new serial port in order to control the serial device correctly.

It is also possible that the serial device may malfunction in an
unrecoverable state. Removing and modprobing the serial drivers again
cannot resolve the problem when the device falls in the state. The only
way is to reboot the chameleon board. In this case, it is required to
reconnect to the remote server at run time. This has been handled in
a transparent way by chameleond_proxy in ChameleonConnection.

BUG=chromium:610238
TEST=conduct the test with the chameleon board connected to an RN-42 kit
(cr) $ test_that --board=$BOARD --args "chameleon_host=$CHAMELEON_IP"
     $DUT_IP bluetooth_AdapterPairing

A user could unplug the RN-42 kit initially. This would lead to the
reboot of the chameleon board. When the board is being rebooted, plug
in the RN-42 kit again to the chameleon board. Observe that the test
could continue smoothly after rebooting the chameleon board.

Change-Id: I43aa81311fbb163bfce6790440c460b2f8426e4c
Reviewed-on: https://chromium-review.googlesource.com/359478
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>