platform_FilePerms: add selinuxfs and android container mount info
This change adds information about selinuxfs and android container
mounts to platform_FilePerms test.

Note that Android container starts as "noexec" but later (once user logs
in) is remounted as "exec", so the test will fail if executed after user
logging in.

Also removed "relatime" from the lists of cheked options as it is the
default since 2.6.30.

BUG=b:29003204
TEST=Booted minnie, ran platform_FilePerms test.

Change-Id: Id7b207325a53152d1921027070012219ff81485d
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/351221
Reviewed-by: Mike Frysinger <vapier@chromium.org>