[autotest] Use a RepairStrategy to repair ServoHost.
This converts ServoHost to use the new repair framework.  It
effectively replaces the now-defunct infra_ServoDiagnosis.

With this change, the preferred way to make sure a Servo is
available and working is simply to invoke repair() on the associated
ServoHost object.  That invocation also assures that standard repair
and verify logging will be present in status.log.

BUG=chromium:630018,chromium:406173,chromium:531643,chromium:543357
TEST=run `repair_test`, run Repair tasks in a local Autotest instance

Change-Id: I5b57e6b4920e043a5389149b3f11dbcd1e59d061
Reviewed-on: https://chromium-review.googlesource.com/372980
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>