[autotest] Simplify the logic of create_servo_host().
This refactors the logic in create_servo_host() with the intent of
separating "get servo arg values" from "decide whether to create a
ServoHost object".  The new code is meant to be easier to understand.
Along the way, it may also be fixing some bugs.

BUG=None
TEST=None

Change-Id: I1ef77f2bb4e50413335c944331577730b4f7a84a
Reviewed-on: https://chromium-review.googlesource.com/351532
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>