Chameleon: Correct the names of EDIDs and explicitly filter unsupported resolutions
The EDIDs of display_Resolution are manually created. They are not
obtained from some real monitor. Their names having "HDMI"/"DP" are
not correct. So change their names to the version of EDID.

Also modify the filtering rule of EDID to only look at the resolution
due to the corresponding pixel clocks are not supported.

BUG=None
TEST=Ran the affected tests on peach and veyron devices.

Change-Id: Id802f5c0e971873d0895ed9c7ee64c1ee630d1b1
Reviewed-on: https://chromium-review.googlesource.com/352574
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Hsu Wei-Cheng <mojahsu@chromium.org>