Improved tests for host mod and create commands.
This is preparation for a refactor of the logic for those commands. For
host mod removed the --ignore_site_file flag to ensure real logic is
always tested. For host create replaced the existing tests with tests
that cover the site logic including all command line arguments.

Remaining changes are to satisfy linter. Any comments on the docstrings
will be welcome. Much of it was inferred from calls I could find in the
codebase.

BUG=None
TEST=Yes, lots of them

Change-Id: Ib882c54318bca0f29f48de0da8c8f2df5e090a0c
Reviewed-on: https://chromium-review.googlesource.com/354552
Commit-Ready: Justin Giorgi <jgiorgi@google.com>
Commit-Ready: Justin Giorgi <jgiorgi@chromium.org>
Tested-by: Justin Giorgi <jgiorgi@google.com>
Tested-by: Justin Giorgi <jgiorgi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>