update_engine: Add retry to the update_engine check status call
Based on the discussion in the bug, there could exist a race condition
between the update_engine start deamon and the update_engine_client
call. Add a retry to check_update_status call to wait for the update_engine
is ready.  And also remove the 2>&1 in several places to show up the logs that
could helpful for debug.

BUG=chromium:588828
TEST=TEST in a test DUT.  Stop update_engine on the DUT, and run the check
status call and the DUT retried.

Change-Id: I214f5029397c25847c4ed52590b00a10bc918227
Reviewed-on: https://chromium-review.googlesource.com/355692
Commit-Ready: Shuqian Zhao <shuqianz@chromium.org>
Tested-by: Shuqian Zhao <shuqianz@chromium.org>
Reviewed-by: Shuqian Zhao <shuqianz@chromium.org>