autotest: use google-api-python-client to 1.5.1; oauth2client 1.5.2
This change allows us to pull in oauth2client 1.5.2, which is in keeping
with what we use in chromite.

We also upgrade to a newer version of google-api-python-client that no
longer ships with its own conflicting version of oauth2client.

BUG=chromium:622877
TEST=import common; import oauth2client; oaut2client.__file__ -> in
autotest site packages, oauth2client.__version__ = 1.5.2

Change-Id: I4c52d07a1a0cb6bbc07dbd154487447b2acdfdaa
Reviewed-on: https://chromium-review.googlesource.com/355721
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>