autotest: change dut status fields slightly
When we add stats to monarch, we already have a field called "host_name"
which refers to the host sending the data (i.e. the drone or shard).
Having "hostname" refer to the DUT is a recipe for confusion.

Rename this to "dut_host_name". Note the underscore here, to be
consistent with monarch's "host_name" convention.

BUG=None
TEST=None

Change-Id: I457aaea4b0734caeb528106bcd7ca519e2e3dfbc
Reviewed-on: https://chromium-review.googlesource.com/356511
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Paul Hobbs <phobbs@google.com>
Commit-Queue: Paul Hobbs <phobbs@google.com>
Tested-by: Paul Hobbs <phobbs@google.com>