[autotest] Scrub out backslashes in host attributes.
A user can use a ',' in the val portion of the host attribute but
they'll need to escape it with two backslashes '\\'.  However when the
attribute gets stored, a backslash will be part of the val so if a user
did the following:

  # atest host mod -i key=val\\,val2\\,val3

The host attribute stored will look like:

  key: val\,val2\,val3

This just scrubs out the backslashes in the value.

BUG=None
TEST=Check that host can add/mod attributes with '\\,' in the val and
doesn't keep the '\\'. Added unit test.

Change-Id: I5f4d7612d5cde51a5bcdaa9261594f86ac15a6be
Reviewed-on: https://chromium-review.googlesource.com/356772
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>