[Autotest] Add mechanism to record kernel input events.
Add a class which maintains state for kernel event recording.
This class uses a subprocess running evtest to record events,
and supports functions querying the gathered data.

Also, add a function to get the timestamp of the last recorded
JavaScript event (as seen by the touch events page).

Not currently used in a test, but I have one in progress which
can successfully use this code.

BUG=chromium:626212
TEST=ran at my desk; was able to record kernel events as expected

Change-Id: Ib19a9f6df3c2e5e9deb4dc13669d2f76ee0bcf04
Reviewed-on: https://chromium-review.googlesource.com/358625
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>