[autotest] Prevent TPM problems from causing Provision failures.
Sometimes, the TPM on a DUT gets into a bad state, such that it can
cause test failures.  Host verification catches and repairs such
problems, but the problem frequently causes provision failures in
the process.

This changes CrosHost.machine_install() to clear the TPM as part of
the standard AU process, which will eliminate all the provisioning
failures, at the cost of papering over whatever is causing the
problem.

BUG=chromium:611121
TEST=test using machine_install on a local instance

Change-Id: I6b44e8c71b99839245c08d25cb101575bc47a803
Reviewed-on: https://chromium-review.googlesource.com/358840
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Shuqian Zhao <shuqianz@chromium.org>