autotest: make sure neither trunksd nor tcsd runs for faft
For TPM2.0 trunksd is run instead of tcsd. So, make sure it is
stopped before using tpmc in FAFT and re-started afterwards.

Before the change rpc_server.py failed to start for tpm2 because
it tried to stop tcsd, which is not present and not known to
initctl if tpm2 is enabled.

BUG=chrome-os-partner:55210
TEST=Verify that rpc_server.py starts on kevin. verify that
     'test_that --board=kevin $IP firmware_FAFTSetup' stops
     trunksd before running and restarts during cleanup on
     kevin.

Change-Id: I3f9f64fb8250838055e43153899ecded36f1ea56
Reviewed-on: https://chromium-review.googlesource.com/361392
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>