platform_FilePerms: add OBB mount info
Set mount info for OBB which remains mounted after a user has
logged in and out. That happens when the test is not run
individually but as part of the smoke test suite where
previous tests have logged in and out. Similar to CL:361310

Test checks that OBB is mounted ro,nosuid,nodev,noexec,mode=755

BUG=chromium:629133
TEST=Ran manually on cyan-cheets machine and VM

Change-Id: I2579f6dc3d7b3865f8005d7bec2b8d7beaea8393
Reviewed-on: https://chromium-review.googlesource.com/361690
Commit-Ready: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>