Reenable logging, warn on logging failure.
Reenables native crash logging, and considers the inability to
enable crash logging as a warning rather than an error.  This
is important because the provision operation calls job_start(),
and depending on the preexisting image on the DUT, the attempt
to enable logging will fail.  Note that job_start() is called
again to start the test case, so logging will be enabled once
the device is provisioned with the image under test.

BUG=chromium:623743
BUG=chromium:627134
TEST=Local test_droid on Brillo: android_DummyTest brillo_CrashLogging
TEST=Local test_droid on Android: android_DummyTest android_CrashLogging
TEST=Lab test_droid on Brillo: android_DummyTest brillo_CrashLogging
TEST=Guado MobLab on Brillo: android_DummyTest brillo_CrashLogging

Change-Id: I1da04f7f2c7c5133ea7befaf152afb4e2166f406
Reviewed-on: https://chromium-review.googlesource.com/362611
Commit-Ready: Bryan Lewandowski <bryanlew@google.com>
Tested-by: Bryan Lewandowski <bryanlew@google.com>
Reviewed-by: Simran Basi <sbasi@chromium.org>